messed up a declaration

+++ src/includes/petsc_prototypes.h
@@ -64,7 +64,7 @@ extern void *cinit_hacapk_struct(int, void *);
 extern void cdeallocate_hacapk_struct(void *);
 extern void cset_hacapk_struct_coord(void *, double *, double *, double *);
 extern void cset_hacapk_eta(void *, double);
-extern void cset_hacapk_inorm(void *, double);
+extern void cset_hacapk_inorm(void *, int);
 extern void cmake_hacapk_struct_hmat(void *, double);
 /* return this rank's number of stored scalars in the assembled HACApk
    H-matrix (dense leaves ndl*ndt, low-rank leaves kt*(ndl+ndt)); the caller
